Log to a rotating file

An extension of AppenderFile that rotates logfiles based on certain conditions. Please refer to the documentation of rotor::rotate() for the meanings of the extra arguments

Fields

age, size, max_backups, fmt, overwrite, compression, backup_dir

Please see rotor::rotate() for the meaning of these arguments (fmt is passed on as format).

cache_backups, set_cache_backups(x)

TRUE or FALSE. If TRUE (the default) the list of backups is cached, if FALSE it is read from disk every time this appender triggers. Caching brings a significant speedup for checking whether to rotate or not based on the age of the last backup, but is only safe if there are no other programs/functions (except this appender) interacting with the backups.

backups

A data.frame containing information on path, file size, etc... on the available backups of file.

See also

Super classes

lgr::Filterable -> lgr::Appender -> lgr::AppenderFile -> AppenderFileRotating

Methods

Public methods

AppenderFileRotating$new()

AppenderFileRotating$append()

AppenderFileRotating$rotate()

AppenderFileRotating$prune()

AppenderFileRotating$set_file()

AppenderFileRotating$set_size()

AppenderFileRotating$set_max_backups()

AppenderFileRotating$set_compression()

AppenderFileRotating$set_create_file()

AppenderFileRotating$set_backup_dir()

AppenderFileRotating$format()

AppenderFileRotating$clone()

Inherited methods

lgr::Filterable$add_filter()

lgr::Filterable$filter()

lgr::Filterable$remove_filter()

lgr::Filterable$set_filters()

lgr::Appender$set_layout()

lgr::Appender$set_threshold()

lgr::AppenderFile$show()

Method new()

Usage

AppenderFileRotating$new(
  file,
  threshold = NA_integer_,
  layout = LayoutFormat$new(),
  filters = NULL,
  size = Inf,
  max_backups = Inf,
  compression = FALSE,
  backup_dir = dirname(file),
  create_file = TRUE
)

Method append()

Usage

AppenderFileRotating$append(event)

Method rotate()

Usage

AppenderFileRotating$rotate(force = FALSE)

Method prune()

Usage

AppenderFileRotating$prune(max_backups = self$max_backups)

Method set_file()

Usage

AppenderFileRotating$set_file(file)

Method set_size()

Usage

AppenderFileRotating$set_size(x)

Method set_max_backups()

Usage

AppenderFileRotating$set_max_backups(x)

Method set_compression()

Usage

AppenderFileRotating$set_compression(x)

Method set_create_file()

Usage

AppenderFileRotating$set_create_file(x)

Method set_backup_dir()

Usage

AppenderFileRotating$set_backup_dir(x)

Method format()

Usage

AppenderFileRotating$format(color = false, ...)

Method clone()

The objects of this class are cloneable with this method.

Usage

AppenderFileRotating$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.