Log to a JSON File

AppenderJson is a shortcut for AppenderFile with LayoutJson, but comes with an extra method show() and an extra active field data to comfortably access the underlying file.

Usage

x <- AppenderFile$new(file, threshold = NA_integer_, layout =
  LayoutFormat$new(), filters = NULL)
    x$add_filter(filter, name = NULL)
x$append(event)
x$filter(event)
x$remove_filter(pos)
x$set_file(file)
x$set_filters(filters)
x$set_layout(layout)
x$set_threshold(level)
    x$destination
x$file
x$filters
x$layout
x$threshold
    

Creating a New Appender

New Appenders are instantiated with <AppenderSubclass>$new(). For the arguments to new() please refer to the section Fields. You can also modify those fields after the Appender has been created with setters in the form of appender$set_<fieldname>(value)

Fields

file, set_file(file)

character scalar. Path to the desired log file. If the file does not exist it will be created

threshold, set_threshold(level)

character or integer scalar. The minimum log level that triggers this logger. See log_levels

layout, set_layout(layout)

a Layout that will be used for formatting the LogEvents passed to this Appender

destination

The output destination of the Appender in human-readable form (mainly for print output)

filters, set_filters(filters)

a list that may contain functions or any R object with a filter() method. These functions must have exactly one argument: event which will get passed the LogEvent when the Filterable's filter() method is invoked. If all of these functions evaluate to TRUE the LogEvent is passed on. Since LogEvents have reference semantics, filters can also be abused to modify them before they are passed on. Look at the source code of with_log_level() or with_log_value() for examples.

data

Get the log recorded by this Appender as a data.frame

threshold, set_threshold(level)

character or integer scalar. The minimum log level that triggers this logger. See log_levels

layout, set_layout(layout)

a Layout that will be used for formatting the LogEvents passed to this Appender

destination

The output destination of the Appender in human-readable form (mainly for print output)

filters, set_filters(filters)

a list that may contain functions or any R object with a filter() method. These functions must have exactly one argument: event which will get passed the LogEvent when the Filterable's filter() method is invoked. If all of these functions evaluate to TRUE the LogEvent is passed on. Since LogEvents have reference semantics, filters can also be abused to modify them before they are passed on. Look at the source code of with_log_level() or with_log_value() for examples.

Methods

show(n, threshold)

Show the last n log entries with a log level bellow threshold. The log entries will be formatted as in the source JSON file

append(event)

Tell the Appender to process a LogEvent event. This method is usually not called by the user, but invoked by a Logger

filter(event)

Determine whether the LogEvent x should be passed on to Appenders (TRUE) or not (FALSE). See also the active binding filters

add_filter(filter, name = NULL), remove_filter(pos)

Add or remove a filter. When adding a filter an optional name can be specified. remove_filter() can remove by position or name (if one was specified)

See also

Examples

tf <- tempfile()
l <- Logger$new("testlogger", appenders = AppenderJson$new(tf), propagate = FALSE)

l$info("A test message")
l$info("A test message %s strings", "with format strings", and = "custom_fields")

l$appenders[[1]]$show()

#> {"level":400,"timestamp":"2019-05-20 11:18:25","logger":"testlogger","caller":"eval","msg":"A test message"}
#> {"level":400,"timestamp":"2019-05-20 11:18:25","logger":"testlogger","caller":"eval","msg":"A test message with format strings strings","and":"custom_fields"}

l$appenders[[1]]$data

#>   level           timestamp     logger caller
#> 1   400 2019-05-20 11:18:25 testlogger   eval
#> 2   400 2019-05-20 11:18:25 testlogger   eval
#>                                          msg           and
#> 1                             A test message          <NA>
#> 2 A test message with format strings strings custom_fields

# cleanup
unlink(tf)