Log to the POSIX System Log

An Appender that writes to the syslog on supported POSIX platforms. Requires the rsyslog package.

Fields

identifier

character scalar. A string identifying the process; if NULL defaults to the logger name

syslog_levels

a named character vector mapping whose names are log levels as understood by rsyslog::syslog() and whose values are lgr log levels (either character or numeric)

a function that takes a vector of lgr log levels as input and returns a character vector of log levels for rsyslog::syslog().

See also

Super classes

lgr::Filterable -> lgr::Appender -> AppenderSyslog

Methods

Public methods

AppenderSyslog$new()

AppenderSyslog$append()

AppenderSyslog$set_syslog_levels()

AppenderSyslog$set_identifier()

Inherited methods

lgr::Filterable$add_filter()

lgr::Filterable$filter()

lgr::Filterable$remove_filter()

lgr::Filterable$set_filters()

lgr::Appender$format()

lgr::Appender$set_layout()

lgr::Appender$set_threshold()

Method new()

Usage

AppenderSyslog$new(
  identifier = NULL,
  threshold = NA_integer_,
  layout = LayoutFormat$new("%m"),
  filters = NULL,
  syslog_levels = c(CRITICAL = "fatal", ERR = "error", WARNING = "warn", INFO = "info",
    DEBUG = "debug", DEBUG = "trace")
)

Method append()

Usage

AppenderSyslog$append(event)

Method set_syslog_levels()

Usage

AppenderSyslog$set_syslog_levels(x)

Method set_identifier()

Usage

AppenderSyslog$set_identifier(x)

Examples

if (requireNamespace("rsyslog", quietly = TRUE)) {
  lg <- get_logger("rsyslog/test")
  lg$add_appender(AppenderSyslog$new(), "syslog")
  lg$info("A test message")

  if (Sys.info()[["sysname"]] == "Linux"){
    system("journalctl -t 'rsyslog/test'")
  }

  invisible(lg$config(NULL))  # cleanup
}

#> INFO  [06:47:02.379] A test message