Abstract class for logging to tabular structures

NOTE: This is an abstract class. Abstract classes cannot be instantiated directly, but are exported for package developers that want to extend lgr - for example by creating their own Appenders or Layouts. Please refer to the see also section for actual implementations of this class.

AppenderTable is extended by Appenders that write to a data source that can be interpreted as tables, (usually a data.frame). Examples are AppenderDbi, AppenderRjdbc and AppenderDt from the lgrExtra package.

See also

Super classes

lgr::Filterable -> lgr::Appender -> AppenderTable

Active bindings

data

character scalar. Contents of the table, parsed to a data.frame.

data

character scalar. Like $data, but returns a data.table instead (requires the data.table package).

Methods

Public methods

AppenderTable$new()

AppenderTable$show()

AppenderTable$format()

Inherited methods

lgr::Filterable$add_filter()

lgr::Filterable$filter()

lgr::Filterable$remove_filter()

lgr::Filterable$set_filters()

lgr::Appender$append()

lgr::Appender$set_layout()

lgr::Appender$set_threshold()

Method new()

Usage

AppenderTable$new(...)

Method show()

Show recent log entries

Usage

AppenderTable$show(threshold = NA_integer_, n = 20L)

Arguments

threshold

an integer or character threshold. Only show events with a log level at or below this threshold.

n

a positive integer scalar. Show at most that many entries

Method format()

Usage

AppenderTable$format(color = FALSE, ...)