Abstract Class for Layouts

Abstract Class for Layouts

Abstract Class for Layouts

Details

Appenders pass LogEvents to a Layout which formats it for output. For the Layouts included in lgr that means turning the LogEvent into a character string.

For each Appender exist one more more possible Layouts, but not every Layout will work with every Appender. See the package lgrExtra for examples for Layouts that return different data types (such as data.frames) and Appenders that can handle them.

See also

Methods

Public methods

Layout$format_event()

Layout$toString()

Layout$clone()

Method format_event()

Format a log event

Usage

Layout$format_event(event)

Arguments

event

a LogEvent

Method toString()

Usage

Layout$toString()

Method clone()

The objects of this class are cloneable with this method.

Usage

Layout$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.