Check if an R Object is a Filter

Returns TRUE for any R object that can be used as a Filter for Loggers or, Appenders:

a function with the single argument event;

an EventFilter R6::R6 object; or

any object with a $filter(event) method.

Note: A Filter must return a scalar TRUE or FALSE, but this property cannot be checked by is_filter().

is_filter(x)

Arguments

x

any R Object

Value

TRUE or FALSE

See also