Log to a date-stamped rotating file

Log to a date-stamped rotating file

Log to a date-stamped rotating file

Details

This is a simpler version of AppenderFileRotatingTime when the timestamps do not need to include sub-day accuracy.

See also

Super classes

lgr::Filterable -> lgr::Appender -> lgr::AppenderFile -> lgr::AppenderFileRotating -> AppenderFileRotatingDate

Methods

Public methods

AppenderFileRotatingDate$new()

AppenderFileRotatingDate$clone()

Inherited methods

lgr::Filterable$add_filter()

lgr::Filterable$filter()

lgr::Filterable$remove_filter()

lgr::Filterable$set_filters()

lgr::Appender$set_layout()

lgr::Appender$set_threshold()

lgr::AppenderFile$show()

lgr::AppenderFileRotating$append()

lgr::AppenderFileRotating$prune()

lgr::AppenderFileRotating$set_backup_dir()

lgr::AppenderFileRotating$set_compression()

lgr::AppenderFileRotating$set_create_file()

lgr::AppenderFileRotating$set_file()

lgr::AppenderFileRotating$set_max_backups()

lgr::AppenderFileRotating$set_size()

lgr::AppenderFileRotating$format()

lgr::AppenderFileRotating$rotate()

lgr::AppenderFileRotating$set_age()

lgr::AppenderFileRotating$set_cache_backups()

lgr::AppenderFileRotating$set_fmt()

lgr::AppenderFileRotating$set_overwrite()

Method new()

Usage

AppenderFileRotatingDate$new(
  file,
  threshold = NA_integer_,
  layout = LayoutFormat$new(),
  filters = NULL,
  age = Inf,
  size = -1,
  max_backups = Inf,
  compression = FALSE,
  backup_dir = dirname(file),
  fmt = "%Y-%m-%d",
  overwrite = FALSE,
  cache_backups = TRUE,
  create_file = NULL
)

Arguments

size, age, max_backups, compression, backup_dir, fmt, overwrite, cache_backups

see rotor::rotate_date() for the meaning of these arguments. Note that fmt corresponds to format (because $format has a special meaning for R6 classes).

Method clone()

The objects of this class are cloneable with this method.

Usage

AppenderFileRotatingDate$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.