Log to a time-stamped rotating file

Log to a time-stamped rotating file

Log to a time-stamped rotating file

See also

Super classes

lgr::Filterable -> lgr::Appender -> lgr::AppenderFile -> lgr::AppenderFileRotating -> AppenderFileRotating

Active bindings

cache_backups

TRUE or FALSE. If TRUE (the default) the list of backups is cached, if FALSE it is read from disk every time this appender triggers. Caching brings a significant speedup for checking whether to rotate or not based on the age of the last backup, but is only safe if there are no other programs/functions (except this appender) interacting with the backups.

Methods

Public methods

AppenderFileRotatingTime$new()

AppenderFileRotatingTime$rotate()

AppenderFileRotatingTime$set_age()

AppenderFileRotatingTime$set_fmt()

AppenderFileRotatingTime$set_overwrite()

AppenderFileRotatingTime$set_cache_backups()

AppenderFileRotatingTime$format()

AppenderFileRotatingTime$clone()

Inherited methods

lgr::Filterable$add_filter()

lgr::Filterable$filter()

lgr::Filterable$remove_filter()

lgr::Filterable$set_filters()

lgr::Appender$set_layout()

lgr::Appender$set_threshold()

lgr::AppenderFile$show()

lgr::AppenderFileRotating$append()

lgr::AppenderFileRotating$prune()

lgr::AppenderFileRotating$set_backup_dir()

lgr::AppenderFileRotating$set_compression()

lgr::AppenderFileRotating$set_create_file()

lgr::AppenderFileRotating$set_file()

lgr::AppenderFileRotating$set_max_backups()

lgr::AppenderFileRotating$set_size()

Method new()

Usage

AppenderFileRotatingTime$new(
  file,
  threshold = NA_integer_,
  layout = LayoutFormat$new(),
  filters = NULL,
  age = Inf,
  size = -1,
  max_backups = Inf,
  compression = FALSE,
  backup_dir = dirname(file),
  fmt = "%Y-%m-%d--%H-%M-%S",
  overwrite = FALSE,
  create_file = TRUE,
  cache_backups = TRUE
)

Method rotate()

Usage

AppenderFileRotatingTime$rotate(force = FALSE, now = Sys.time())

Method set_age()

Usage

AppenderFileRotatingTime$set_age(x)

Method set_fmt()

Usage

AppenderFileRotatingTime$set_fmt(x)

Method set_overwrite()

Usage

AppenderFileRotatingTime$set_overwrite(x)

Method set_cache_backups()

set the cache_backups flag.

Usage

AppenderFileRotatingTime$set_cache_backups(x)

Arguments

x

a logical scalar

Method format()

Usage

AppenderFileRotatingTime$format(color = FALSE, ...)

Method clone()

The objects of this class are cloneable with this method.

Usage

AppenderFileRotatingTime$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.