Abstract class for logging to memory buffers

This is an abstract class and as such cannot be instantiated directly. It is exported for package developers that want to extend it

AppenderMemory is extended by Appenders that retain an in-memory event buffer, such as AppenderBuffer and lgrExtra::AppenderPushbullet.

Fields

buffer_size, set_buffer_size(x)

integer scalar >= 0 Number of LogEvents to buffer.

buffer_events, buffer_df, buffer_dt

The contents of the buffer as a list of LogEvents, a data.frame or a data.table.

flush_threshold, set_flush_threshold()

integer or character log level. Minimum event level that will trigger flushing of the buffer (set NULL for no flushing). This behaviour is implemented through should_flush(), and you can modify that function for different behaviour.

should_flush(event), set_should_flush(x)

A function with exactly one arguments: event. If the function returns TRUE, flushing of the buffer is triggered. Defaults to flushing if an event of level error or higher is registered.

See also

Super classes

lgr::Filterable -> lgr::Appender -> AppenderMemory

Methods

Public methods

AppenderMemory$new()

AppenderMemory$append()

AppenderMemory$flush()

AppenderMemory$clear()

AppenderMemory$set_buffer_size()

AppenderMemory$set_should_flush()

AppenderMemory$set_flush_on_exit()

AppenderMemory$set_flush_on_rotate()

AppenderMemory$set_flush_threshold()

AppenderMemory$show()

AppenderMemory$format()

Inherited methods

lgr::Filterable$add_filter()

lgr::Filterable$filter()

lgr::Filterable$remove_filter()

lgr::Filterable$set_filters()

lgr::Appender$set_layout()

lgr::Appender$set_threshold()

Method new()

Usage

AppenderMemory$new(...)

Method append()

Usage

AppenderMemory$append(event)

Method flush()

Sends the buffer's contents to all attached Appenders and then clears the Buffer

Usage

AppenderMemory$flush()

Method clear()

Clears the buffer, discarding all buffered Events

Usage

AppenderMemory$clear(self)

Method set_buffer_size()

Usage

AppenderMemory$set_buffer_size(x)

Method set_should_flush()

Usage

AppenderMemory$set_should_flush(x)

Method set_flush_on_exit()

Usage

AppenderMemory$set_flush_on_exit(x)

Method set_flush_on_rotate()

Usage

AppenderMemory$set_flush_on_rotate(x)

Method set_flush_threshold()

Usage

AppenderMemory$set_flush_threshold(level)

Method show()

Usage

AppenderMemory$show(threshold = NA_integer_, n = 20L)

Method format()

Usage

AppenderMemory$format(color = FALSE, ...)