Abstract class for logging to memory buffers

NOTE: This is an abstract class. Abstract classes cannot be instantiated directly, but are exported for package developers that want to extend lgr - for example by creating their own Appenders or Layouts. Please refer to the see also section for actual implementations of this class.

AppenderMemory is extended by Appenders that retain an in-memory event buffer, such as AppenderBuffer and AppenderPushbullet from the lgrExtra package.

See also

Super classes

lgr::Filterable -> lgr::Appender -> AppenderMemory

Active bindings

flush_on_exit

A logical scalar. Should the buffer be flushed if the Appender is destroyed (e.g. because the R session is terminated)?

flush_on_rotate

A logical scalar. Should the buffer be flushed when it is rotated because $buffer_size is exceeded?

should_flush

A function with exactly one arguments: event. $append() calls this function internally on the current LogEvent and flushes the buffer if it evaluates to TRUE.

buffer_size

integer scalar >= 0. Maximum number of LogEvents to buffer.

flush_threshold

A numeric or character threshold. LogEvents with a log_level equal to or lower than this threshold trigger flushing the buffer.

buffer_events

A list of LogEvents. Contents of the buffer.

buffer_events

A data.frame. Contents of the buffer converted to a data.frame.

buffer_events

A data.frame. Contents of the buffer converted to a data.table.

Methods

Public methods

AppenderMemory$new()

AppenderMemory$append()

AppenderMemory$flush()

AppenderMemory$clear()

AppenderMemory$set_buffer_size()

AppenderMemory$set_should_flush()

AppenderMemory$set_flush_on_exit()

AppenderMemory$set_flush_on_rotate()

AppenderMemory$set_flush_threshold()

AppenderMemory$show()

AppenderMemory$format()

Inherited methods

lgr::Filterable$add_filter()

lgr::Filterable$filter()

lgr::Filterable$remove_filter()

lgr::Filterable$set_filters()

lgr::Appender$set_layout()

lgr::Appender$set_threshold()

Method new()

Usage

AppenderMemory$new(...)

Method append()

Usage

AppenderMemory$append(event)

Method flush()

Sends the buffer's contents to all attached Appenders and then clears the Buffer

Usage

AppenderMemory$flush()

Method clear()

Clears the buffer, discarding all buffered Events

Usage

AppenderMemory$clear()

Method set_buffer_size()

Set the maximum size of the buffer

Usage

AppenderMemory$set_buffer_size(x)

Arguments

x

an integer scalar >= 0. Number of LogEvents to buffer.

Method set_should_flush()

Set function that can trigger flushing the buffer

Usage

AppenderMemory$set_should_flush(x)

Arguments

x

A function with the single argument event. Setting x to NULL is a shortcut for function(event) FALSE. See active bindings.

Method set_flush_on_exit()

Should the buffer be flushed when the Appender is destroyed?

Usage

AppenderMemory$set_flush_on_exit(x)

Arguments

x

A logical scalar. See active bindings.

Method set_flush_on_rotate()

Should the buffer be flushed if buffer_size is exceeded?

Usage

AppenderMemory$set_flush_on_rotate(x)

Arguments

x

A logical scalar. See active bindings.

Method set_flush_threshold()

Set threshold that triggers flushing

Usage

AppenderMemory$set_flush_threshold(level)

Arguments

level

A numeric or character threshold. See active bindings.

Method show()

Display the contents of the log table. Relies on the $format_event method of the Layout attached to this Appender.

Usage

AppenderMemory$show(threshold = NA_integer_, n = 20L)

Arguments

threshold

character or integer scalar. The minimum log level that should be displayed.

n

integer scalar. Show only the last n log entries that match threshold.

Method format()

Usage

AppenderMemory$format(color = FALSE, ...)