Event Filters

Filters can be used for the $set_filter() and $add_filter() methods of Appenders and Loggers. You normally do not need to construct a formal EventFilter object, you can just use any function that has the single argument event, or any object that has a filter method.

NOTE: You can use the special function .obj() to access the calling Logger/Appender from within a filter

Modifying LogEvents with Filters

Since LogEvents are R6 objects with reference semantics, Filters can be abused to modify log events before passing them on. lgr comes with a few preset filters that use this property: FilterInject (similar to with_log_level()) and FilterForceLevel (similar to with_log_value()).

The base class for Filters is called EventFilter so that it doesn't conflict with base::Filter(). The recommended convention for Filter subclasses is to call them FilterSomething and leave out the Event prefix.

Methods

Public methods

EventFilter$new()

EventFilter$clone()

Method new()

Initilize a new EventFilter

Usage

EventFilter$new(fun = function(event) TRUE)

Arguments

fun

a function with a single argument event that must return either TRUE or FALSE.

Method clone()

The objects of this class are cloneable with this method.

Usage

EventFilter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.