Abstract Class for Filterables

This is an abstract class and as such cannot be instantiated directly. It is exported for package developers that want to extend it

Superclass for classes that have a $filter() method such as Appenders and Loggers. See EventFilter for Details.

See also

Active bindings

filters

a list of all attached filters (either Filter R6 objects or functions).

Methods

Public methods

Filterable$filter()

Filterable$add_filter()

Filterable$remove_filter()

Filterable$set_filters()

Method filter()

Determine whether the LogEvent x should be passed on to Appenders (TRUE) or not (FALSE). See also the active binding filters

Usage

Filterable$filter(event)

Arguments

event

a LogEvent

Method add_filter()

Add a Filter. When adding a filter an optional name can be specified. remove_filter() can remove by position or name (if one was specified)

Usage

Filterable$add_filter(filter, name = NULL)

Arguments

filter

a function that returns TRUE or FALSE, a Filter or any R object with a $filter() method.

name

character scalar or NULL. A filter can have an optional name which makes it easier to access (or remove) the filter

Method remove_filter()

Remove a filter

Usage

Filterable$remove_filter(pos)

Arguments

pos

character or integer scalar. The name or index of the Filter to be removed.

Method set_filters()

Replace all filters with a list of either functions or arbitrary R object with a $filter() method (preferably a Filter R6 object). These functions/methods must have exactly one argument: event which will get passed the LogEvent when the Filterable's $filter() method is invoked. If all of these functions evaluate to TRUE the LogEvent is passed on. Since LogEvents have reference semantics, filters can also be abused to modify them before they are passed on. Look at the source code of with_log_level() or with_log_value() for examples.

Usage

Filterable$set_filters(filters)

Arguments

filters

a list (named or unnamed) of Filter or predicate functions.