Format Log Events for Output to Databases

DBI Layouts convert LogEvents to a data.frames that can be appended to an existing database table. Via the col_types parameter they can also store information that AppenderDbi can use to create new database tables; however, it is safer and more flexible to set up the log table up manually with an SQL CREATE TABLE statement instead.

Usage

x <- x$new(event_values = NULL, col_types = NULL)
    x$format_event(event)
x$set_col_types(x)
x$sql_create_table(table)
x$clone(deep = FALSE)
x$set_event_values(x)
x$set_val_names(x)
    x$col_names
x$col_types
x$event_values
x$val_names
    

Creating a New Layout

Layouts are instantiated with <LayoutSubclass>$new(). For a description of the arguments to this function please refer to the Fields section.

Fields

col_types

A named character vector of column types supported by the target database. If this is used instead of event_values, and the target logging table does not yet exist, the column type information is used by AppenderDbi or similar Appenders to create a new database table, either on instantion of the Appender or on writing of the first LogEvent. If the target database table already exists, the column type information is not used. You can only supply one of event_values and col_types.

col_names

Convenience method to get the names of the col_types vector

Methods

format_event(event)

format a LogEvent

Database Specific Layouts

Different databases have different data types and features. Currently the following LayoutDBI subclasses exist that deal with specific databases, but this list is expected to grow as lgr matures:

LayoutSQLite: Needs its own Layout because SQLite does not support timestamps

LayoutDBI: For all other datbases

The utility function select_dbi_layout() returns the appropriate Layout for a DBI connection.

See also

Examples

# setup a dummy LogEvent
event <- LogEvent$new(
  logger = Logger$new("dummy logger", user = "testuser"),
  level = 200,
  timestamp = Sys.time(),
  caller = NA_character_,
  msg = "a test message"
)

# defaults
lo <- LayoutDbi$new()
lo$format_event(event)

#>   level           timestamp caller            msg
#> 1   200 2019-01-03 16:16:49   <NA> a test message

# SQLite does not support timestamps so LayoutSqlite converts them to text
lo <- LayoutSqlite$new()
str(lo$format_event(event))

#> 'data.frame':	1 obs. of  4 variables:
#>  $ level    : num 200
#>  $ timestamp: chr "2019-01-03 16:16:49"
#>  $ caller   : chr NA
#>  $ msg      : chr "a test message"

# advanced example that supports a custom_field:
lo <- LayoutDbi$new(
  col_types =  c(
    timestamp = "timestamp",
    level = "smallint",
    msg = "varchar(2048)",
    custom_field = "integer"
  )
)

event <- LogEvent$new(
  logger = Logger$new("dummy logger", user = "testuser"),
  level = 200,
  timestamp = Sys.time(),
  caller = NA_character_,
  msg = "a test message",
  custom_field = "blubb"
)

lo$format_event(event)

#>             timestamp level            msg custom_field
#> 1 2019-01-03 16:16:49   200 a test message        blubb