Format Log Events as Text via glue

Format a LogEvent as human readable text using glue::glue. The function is evaluated in an environment in which it has access to all elements of the LogEvent (see examples). This is more flexible than LayoutFormat, but also more complex and slightly less performant.

Usage

x <- LayoutGlue$new(fmt = "{pad_right(colorize_levels(toupper(level_name)), 5)}
  [{timestamp}] msg")

x$clone(deep = FALSE)
x$format_event(event)
x$set_colors(x)
x$set_fmt(x)
x$toString()

x$fmt

Creating a New LayoutGlue

A new LayoutGlue is instantiated with LayoutGlue$new(). It takes a single argument fmt that is passed on to glue::glue() for each LogEvent.

Fields

fmt

see glue::glue()

Methods

format_event(event)

format a LogEvent

See also

Examples

lg <- get_logger("test")$
  set_appenders(AppenderConsole$new())$
  set_propagate(FALSE)

lg$appenders[[1]]$set_layout(LayoutGlue$new())
lg$fatal("test")

#> FATAL [2019-09-11 12:25:42] msg

# All fields of the LogEvent are available, even custom ones
lg$appenders[[1]]$layout$set_fmt(
  "{logger$name} {level_name}({level}) {caller}: {toupper(msg)} {{custom: {custom}}}"
)
lg$fatal("test", custom = "foobar")

#> Warning: [2019-09-11 12:25:42.181] An error occurred during logging: Error in logger$name: $ operator is invalid for atomic vectors

lg$config(NULL)  # reset logger config

#> <Logger> [info] test
#> 
#> inherited appenders:
#>   console: <AppenderConsole> [all] -> console