Format LogEvents as JSON

Format LogEvents as JSON

Format LogEvents as JSON

See also

Super class

lgr::Layout -> LayoutJson

Active bindings

toJSON_args

a list of values passed on to jsonlite::toJSON()

timestamp_fmt

Used by $format_event() to format timestamps.

Methods

Public methods

LayoutJson$new()

LayoutJson$format_event()

LayoutJson$set_toJSON_args()

LayoutJson$set_timestamp_fmt()

LayoutJson$toString()

LayoutJson$parse()

LayoutJson$read()

LayoutJson$clone()

Inherited methods

Method new()

Usage

LayoutJson$new(toJSON_args = list(auto_unbox = TRUE), timestamp_fmt = NULL)

Method format_event()

Usage

LayoutJson$format_event(event)

Method set_toJSON_args()

Set a format that this Layout will apply to timstamps. can be a character scalar as for format.POSIXct() or a function.

Usage

LayoutJson$set_toJSON_args(x)

Method set_timestamp_fmt()

Set a format that this Layout will apply to timstamps.

Usage

LayoutJson$set_timestamp_fmt(x)

Arguments

x

NULL (the default): formatting of the timestamp is left to jsonlite::toJSON(),

a character scalar as for format.POSIXct(), or

a function that must return a vector of the same length as its input. The returned vector can be of any type supported by jsonlite::toJSON(), but usually it should be character.

Method toString()

Usage

LayoutJson$toString()

Method parse()

Usage

LayoutJson$parse(file)

Method read()

Usage

LayoutJson$read(file, threshold = NA_integer_, n = 20L)

Method clone()

The objects of this class are cloneable with this method.

Usage

LayoutJson$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# setup a dummy LogEvent

event <- LogEvent$new(
  logger = Logger$new("dummy logger"),
  level = 200,
  timestamp = Sys.time(),
  caller = NA_character_,
  msg = "a test message",
  custom_field = "LayoutJson can handle arbitrary fields"
)

# Default settings show all event fals
lo <- LayoutJson$new()
lo$format_event(event)

#> {"level":200,"timestamp":"2020-03-20 21:08:35","logger":"dummy logger","caller":null,"msg":"a test message","custom_field":"LayoutJson can handle arbitrary fields"}