LoggerGlue

LoggerGlue uses glue::glue() instead of base::sprintf() to construct log messages. glue is a very well designed package for string interpolation. It makes composing log messages more flexible and comfortable at the price of an additional dependency and slightly less performance than sprintf().

Details

glue() lets you define temporary named variables inside the call. As with the normal Logger, these named arguments get turned into custom fields; however, you can suppress this behaviour by making named argument start with a ".". Please refer to vignette("lgr", package = "lgr") for examples.

Super classes

lgr::Filterable -> lgr::Logger -> LoggerGlue

Methods

Public methods

LoggerGlue$fatal()

LoggerGlue$error()

LoggerGlue$warn()

LoggerGlue$info()

LoggerGlue$debug()

LoggerGlue$trace()

LoggerGlue$log()

LoggerGlue$list_log()

LoggerGlue$spawn()

Inherited methods

lgr::Filterable$add_filter()

lgr::Filterable$filter()

lgr::Filterable$remove_filter()

lgr::Filterable$set_filters()

lgr::Logger$add_appender()

lgr::Logger$config()

lgr::Logger$handle_exception()

lgr::Logger$initialize()

lgr::Logger$remove_appender()

lgr::Logger$set_appenders()

lgr::Logger$set_exception_handler()

lgr::Logger$set_propagate()

lgr::Logger$set_threshold()

Method fatal()

Usage

LoggerGlue$fatal(..., caller = get_caller(-8L), .envir = parent.frame())

Method error()

Usage

LoggerGlue$error(..., caller = get_caller(-8L), .envir = parent.frame())

Method warn()

Usage

LoggerGlue$warn(..., caller = get_caller(-8L), .envir = parent.frame())

Method info()

Usage

LoggerGlue$info(..., caller = get_caller(-8L), .envir = parent.frame())

Method debug()

Usage

LoggerGlue$debug(..., caller = get_caller(-8L), .envir = parent.frame())

Method trace()

Usage

LoggerGlue$trace(..., caller = get_caller(-8L), .envir = parent.frame())

Method log()

Usage

LoggerGlue$log(
  level,
  ...,
  timestamp = Sys.time(),
  caller = get_caller(-7),
  .envir = parent.frame()
)

Method list_log()

Usage

LoggerGlue$list_log(x)

Method spawn()

Usage

LoggerGlue$spawn(name)