Logger Configuration Objects

logger_config() is an S3 constructor for logger_config objects that can be passed to the $config method of a Logger. You can just pass a normal list instead, but using this constructor is a more formal way that includes additional argument checking.

logger_config(appenders = list(), threshold = NULL, filters = list(),
  exception_handler = default_exception_handler, propagate = TRUE)

as_logger_config(x)

# S3 method for list
as_logger_config(x)

# S3 method for character
as_logger_config(x)

# S3 method for Logger
as_logger_config(x)

Arguments

appenders
see Logger

threshold
see Logger

filters
see Logger

exception_handler
see Logger

propagate
see Logger

x
any R object. Especially:

A character scalar. This can either be the path to a YAML file or a character scalar containing valid YAML

a list containing the elements appenders, threshold, exception_handler, propagate and filters. See the section Fields in Logger for details.

a Logger object, to clone its configuration.

Value

a list with the subclass "logger_config"

a logger_config object

See also

Examples

lg <- get_logger("test")

# explicetely defining logger configurations with logger_config

# call without arguments to generate the default configuration
cfg <- logger_config()  # same as the unconfigured state of a Logger
lg$config(cfg)

#> <Logger> [all] test
#> 
#> inherited appenders:
#>   console: <AppenderConsole> [info] -> console
#>   memory : <AppenderDt>      [ all] -> in memory data.table

# Creating a logger config form YAML
cfg <- "
Logger:
  name: test/blubb
  threshold: info
  propagate: false
  appenders:
    AppenderFile:
      file: /tmp/testlog.txt
"
lg$config(cfg)  # calls as_logger_config() internally

#> <Logger> [info] test
#> 
#> appenders:
#>   [[1]]: <AppenderFile> [all] -> /tmp/testlog.txt

lg$config(logger_config())  # reset logger config to default state

#> <Logger> [all] test
#> 
#> inherited appenders:
#>   console: <AppenderConsole> [info] -> console
#>   memory : <AppenderDt>      [ all] -> in memory data.table