Print a Logger Object

The print() method for Loggers displays the most important aspects of the Logger.

You can also print just the ancestry of a Logger which can be accessed with with logger$ancestry(). This returns a named character vector whose names correspond to the names of the Loggers logger inherits from. The TRUE/FALSE status of its elements correspond to the propagate values of these Loggers.

# S3 method for Logger
print(x, color = requireNamespace("crayon", quietly = TRUE), ...)

# S3 method for Logger
format(x, color = FALSE, ...)

# S3 method for ancestry
print(x, color = requireNamespace("crayon", quietly = TRUE), ...)

# S3 method for ancestry
format(x, color = FALSE, ...)

Arguments

x
any R Object

color
TRUE or FALSE: Output with color? Requi<- the Package crayon

...
ignored

Value

print() returns x (invisibly), format() returns a character vector.

Examples

# print most important details of logger
print(lgr)

#> <LoggerRoot> [info] root
#> 
#> appenders:
#>   console: <AppenderConsole> [all] -> console

# print only the ancestry of a logger
lg <- get_logger("AegonV/Aerys/Rheagar/Aegon")
get_logger("AegonV/Aerys/Rheagar")$set_propagate(FALSE)

print(lg$ancestry)

#> AegonV/Aerys/Rheagar/Aegon 

unclass(lg$ancestry)

#>  AegonV   Aerys Rheagar   Aegon 
#>    TRUE    TRUE   FALSE    TRUE